% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ensembl.R
\name{annotate_ensembl}
\alias{annotate_ensembl}
\title{Takes an ensembl_gene_id}
\usage{
annotate_ensembl(IDs)
}
\arguments{
\item{IDs}{All IDs (rat or human); all are ensembl_gene_id}
}
\value{
The object resultsContrast annotated with an associated color for the selected comparison
}
\description{
Takes an ensembl_gene_id
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{colorNumericValues}} #Hyperlink does not seem to be working...
}
\keyword{IDs}
\keyword{ensembl}

