% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionsChoice.R
\name{conditionsChoice}
\alias{conditionsChoice}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
conditionsChoice(comparisonSelection, names)
}
\arguments{
\item{comparisonSelection}{Selected comparison of 2 groups}
}
\value{
A vector of colors. The darkest green values are the lowest and the darkest red values are the highest.
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
comparisonSelection <- "DORSAL\\\\VENTRAL"
names <- get_names(expr.matrix)
conditionsChoice(comparisonSelection,names)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[GEOquery]{getGEO}}
\code{\link[Biobase]{ExpressionSet}}
}
\keyword{comparisons}

