% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_contrasts.R
\name{design_contrasts}
\alias{design_contrasts}
\title{Create a design matrix for linear models analysis}
\usage{
design_contrasts(names)
}
\arguments{
\item{names}{A vector of the names of all samples}
}
\value{
a design matrix for linear models analysis
}
\description{
Create a design matrix for linear models analysis
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
design <- design_contrasts(expr.matrix)
lm2 <- lm2Contrast(expr.matrix,design)
lm2.contrast = lm2[[1]]
contrasts=lm2[[2]]
contrast.matrix=lm2[[3]]
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{design_contrasts}}
\code{\link{makeContrasts}}
\code{\link[limma]{lmFit}}
}
\keyword{design}
\keyword{limma}
\keyword{linear}

