% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2cytoscape.R
\name{export2cytoscape}
\alias{export2cytoscape}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
export2cytoscape(expr.matrix, results, threshold = 0.9,
  typeID = "ensembl_gene_id")
}
\arguments{
\item{expr.matrix}{A vector of the names of all samples}

\item{results}{A list of matrices. Each matrix contain a}

\item{threshold}{The lowest value of correlation to make the link
considered significantly high}

\item{typeID}{The type of ID used, e.g.}
}
\value{
A vector of all possible comparisons
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\examples{
expr.matrix
results
networkSelectedComparison = export2cytoscape(expr.matrix,results[[selectedComparison]],threshold,typeID)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{}}
}
\keyword{cytoscape}
\keyword{export}

