% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get names of each sample and group names}
\usage{
get_names(names)
}
\arguments{
\item{names}{Samples names}
}
\value{
return two objects:names1 -> all names without the number
(if their was one) which makes the name of samples in the same groups identical.
names.unique -> A vector of the name of each group (remove identical names from names1)
}
\description{
Get names of each sample and group names
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
pval <- 0.05
summaryClust <- clusteringSummary(expr.matrix,pval)
}
\author{
Simon J Pelletier
}
\keyword{names}
\keyword{names.unique}

