% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMatrix.R
\name{meanMatrix}
\alias{meanMatrix}
\title{Mean matrix}
\usage{
meanMatrix(expr.matrix.signif)
}
\value{
Matrix
\describe{
 \item{columns}{Groups (mean values)}
 \item{rows}{Genes}
}
}
\description{
For every row of the initial expression matrix, the mean for each group is calculated.
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
results <- readRDS("data/results_LGVD.rds")
topTable3 <- readRDS("data/results_LGVD.rds")
resultsSummary <- results_summary(results,topTable3,adjust="BH",typeID="ensembl_gene_id")
expr.matrix.signif <- signifRows(expr.matrix,resultsSummary)

}
\seealso{
\code{\link{signifRows}}
}

