% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_summary.R
\name{modules_summary}
\alias{modules_summary}
\title{Generate boxplots of every groups for a selected gene. Also display stars for significance between groups}
\usage{
modules_summary(results, expr.toBind)
}
\arguments{
\item{results}{The results of one comparison
between at least two groups. Columns = samples, rows = genes,transcripts,CpG...}
}
\description{
Generate boxplots of every groups for a selected gene. Also display stars for significance between groups
}
\examples{
results <- readRDS("data/results_LGVD.rds")
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
bnet <- readRDS("data/bnet_LGVD.rds")
expr.toBind <- cbind(module=bnet$colors,ensembl_gene_id=rownames(expr.matrix),expr.matrix)
modulesTable <- modules_summary(results,expr.toBind)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[stats]{phyper}}
}
\keyword{annotation}
\keyword{geo}

