% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.graphite.R
\name{network.graphite}
\alias{network.graphite}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
network.graphite(pSymbol, expr.matrix, toptableNodes, typeID, threshold,
  selectedReaction, toptable3, selectedComparison, expr.toBind, pvalue)
}
\arguments{
\item{pSymbol}{.}

\item{expr.matrix}{Expression set containing all the information on a dataset}

\item{typeID}{Type of IDs used as rows in the expression matrix.Default="ensembl_gene_id"}

\item{pvalue}{p-value limit}

\item{topTable3}{topTable results for every gene}
}
\value{
A vector of colors. The darkest green values are the lowest and the darkest red values are the highest.
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
topReactions <- readRDS("data/reactions_mfuzz.rds")
topTable3 <- readRDS("data/topTable3_LGVD.rds")
selectedReaction <- "RHO GTPases activate PAKs"
ratReactome <- pathways("rnorvegicus", "reactome")
p <- ratReactome[[selectedReaction]]
pSymbol <- convertIdentifiers(p, "SYMBOL")
toptableNodes = cbind(
   topTable3[[selectedComparison]][match(nodes(pSymbol),topTable3[[selectedComparison]]$symbol),],
   Reaction = rep(selectedReaction[[1]],length(nodes(pSymbol)))
)
toptableNodes1 = toptableNodes[!is.na(toptableNodes[,1]),]
typeID = "ensembl_gene_id"
threshold = 0.8

network.graphite(pSymbol,expr.matrix,toptableNodes,typeID,threshold,selectedReaction,toptable3,selectedComparison,expr.toBind,pvalue)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[GEOquery]{getGEO}}
\code{\link[Biobase]{ExpressionSet}}
}
\keyword{comparisons}

