% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibleComparisons.R
\name{possibleComparisons}
\alias{possibleComparisons}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
possibleComparisons(names)
}
\arguments{
\item{names}{A vector of the names of all samples}
}
\value{
A vector of all possible comparisons
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\author{
Simon J Pelletier
}
\keyword{comparisons}

