% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_topTable.R
\name{results_topTable}
\alias{results_topTable}
\title{Calculate results}
\usage{
results_topTable(lm2.contrast, expr.toBind, pvalue, logFC, typeID,
  genes_annotation_unique, annotations, adjust)
}
\arguments{
\item{lm2.contrast}{.}

\item{expr.toBind}{.}

\item{pvalue}{.}

\item{logFC}{.}

\item{typeID}{.}

\item{genes_annotation_unique}{.}

\item{annotations}{.}

\item{adjust}{.}
}
\value{
Results .
}
\description{
Calculate results
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
design <- design_contrasts(expr.matrix)
results_list <- results_topTable()
lm2 <- lm2Contrast(expr.matrix,design)
lm2.contrast = lm2[[1]]
contrasts=lm2[[2]]
contrast.matrix=lm2[[3]]
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}}
\code{\link{transcript_count}}
}
\keyword{limma}
\keyword{linear}

