% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_modules.R
\name{WGCNA_modules}
\alias{WGCNA_modules}
\title{Find to which genes belong to the same modules}
\usage{
WGCNA_modules(expr.matrix, ncores = 1, names.unique)
}
\arguments{
\item{expr.matrix}{A matrix of expression values. Rows are genes, columns are samples}

\item{ncores}{Number of cores available for parallel programming (foreach function)}

\item{names.unique}{Names of all groups of sample}
}
\value{
A list of WGCNA results
}
\description{
Find to which genes belong to the same modules
}
\examples{
ONLY USE A SUBSET; CAN TAKE A VERY LONG TIME TO RUN
results
networkSelectedComparison = export2cytoscape(expr.matrix,results[[selectedComparison]],threshold,typeID)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{WGCNA}}
}
\keyword{cytoscape}
\keyword{export}

