% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_summary.R
\name{modules_summary}
\alias{modules_summary}
\title{Returns the summary for a module}
\usage{
modules_summary(results, expr.toBind)
}
\arguments{
\item{results}{The results of a comparison between two groups.}

\item{expr.toBind}{Expression data.matrix with supplemental colums: module and ID}
}
\description{
Returns the summary for a module
}
\examples{
results <- readRDS("data/results_LGVD.rds")
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
bnet <- readRDS("data/bnet_LGVD.rds")
expr.toBind <- cbind(module=bnet$colors,ensembl_gene_id=rownames(expr.matrix),expr.matrix)
modulesTable <- modules_summary(results,expr.toBind)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[stats]{phyper}}
}
\keyword{annotation}
\keyword{geo}

