% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{results_summary}
\alias{results_summary}
\title{Summary of the results
Create a data.frame that displays every gene significant in any comparison and
the p-value for each comparison.}
\usage{
results_summary(results, topTable3, adjust = "no",
  typeID = "ensembl_gene_id")
}
\arguments{
\item{expr.matrix}{A matrix of standardized data. Columns = samples, rows = genes,transcripts,CpG...}

\item{resultsSummary}{A matrix of data from every gene that is significanlty different}
}
\value{
The significant rows of the initial expression matrix
between at least two groups. Columns = samples, rows = genes,transcripts,CpG...
}
\description{
Summary of the results
Create a data.frame that displays every gene significant in any comparison and
the p-value for each comparison.
}
\examples{
results <- readRDS("data/results_LGVD.rds")
topTable3 <- readRDS("data/topTable3_LGVD.rds")
resultsSummary <- results_summary(results,topTable3)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}}
}

